requirements.txt
openai>=1.0.0
openpyxl>=3.1.0
pypdf>=5.0.0
python-docx>=1.1.0
reportlab>=4.0.0
